requirements.txt
appdirs==1.4.4
argcomplete==3.6.2
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.2
gunicorn==23.0.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
pyserial==3.5
python-barcode==0.16.1
python-escpos==3.1
pyusb==1.3.1
PyYAML==6.0.2
qrcode==8.2
setuptools==80.9.0
six==1.17.0
Werkzeug==3.1.3
